perf: migrate to Intellij Platform Coroutine Scopes
